CAAEclipses_CalculateLunar

CAAEclipses_CalculateLunar

CAAEclipses_CalculateLunar(k)

Arguments

k
k The same K term as returned from CAAMoonPhases::K. For a lunar eclipse, this value should be decimal value incremented by 0.5 as a lunar eclipse refers to a Full Moon.

Details

Value

A struct containing the following values:

bEclipse true if a lunar eclipse occurs at this Full Moon.

TimeOfMaximumEclipse The date in Dynamical time of maximum eclipse.

F The moons argument of Latitude in degrees at the time of the eclipse.

u The U term for the eclipse.

gamma The gamma term for the eclipse.

PenumbralRadii The radii of the eclipse for the penumbra in equatorial earth radii.

UmbralRadii The radii of the eclipse for the umbra in equatorial earth radii.

PenumbralMagnitude The magnitude of the eclipse if the eclipse is penumbral.

UmbralMagnitude The magnitude of the eclipse if the eclipse is umbral.

PartialPhaseSemiDuration The semi-duration of the eclipse during the partial phase.

TotalPhaseSemiDuration The semi-duration of the eclipse during the total phase.

PartialPhasePenumbralSemiDuration The semi-duration of the partial phase in the penumbra.

References

Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.

Author

C++ code by PJ Naughter, imported to R by Jinlong Zhang

Note

See also

Examples

CAAEclipses_CalculateLunar(k = 2)

#> $bEclipse
#> [1] FALSE
#> 
#> $F
#> [1] 222.051805675786
#> 
#> $gamma
#> [1] 0
#> 
#> $TimeOfMaximumEclipse
#> [1] 0
#> 
#> $u
#> [1] 0
#> 
#> $PenumbralRadii
#> [1] 0
#> 
#> $UmbralRadii
#> [1] 0
#> 
#> $PenumbralMagnitude
#> [1] 0
#> 
#> $UmbralMagnitude
#> [1] 0
#> 
#> $PartialPhaseSemiDuration
#> [1] 0
#> 
#> $TotalPhaseSemiDuration
#> [1] 0
#> 
#> $PartialPhasePenumbraSemiDuration
#> [1] 0
#> 

CAAEclipses_CalculateLunar(k = 24)

#> $bEclipse
#> [1] TRUE
#> 
#> $F
#> [1] 176.80286755311
#> 
#> $gamma
#> [1] 0.408914216164756
#> 
#> $TimeOfMaximumEclipse
#> [1] 2452258.37041329
#> 
#> $u
#> [1] 0.0133398856507003
#> 
#> $PenumbralRadii
#> [1] 1.2981398856507
#> 
#> $UmbralRadii
#> [1] 0.7269601143493
#> 
#> $PenumbralMagnitude
#> [1] 2.13160673300173
#> 
#> $UmbralMagnitude
#> [1] 1.08357045538448
#> 
#> $PartialPhaseSemiDuration
#> [1] 101.698363101123
#> 
#> $TotalPhaseSemiDuration
#> [1] 22.1132352035166
#> 
#> $PartialPhasePenumbraSemiDuration
#> [1] 169.107770652151
#> 

CAAEclipses_CalculateLunar(k = 1)

#> $bEclipse
#> [1] TRUE
#> 
#> $F
#> [1] 191.381302838946
#> 
#> $gamma
#> [1] -1.22322884310893
#> 
#> $TimeOfMaximumEclipse
#> [1] 2451580.03523039
#> 
#> $u
#> [1] 0.02220084350011
#> 
#> $PenumbralRadii
#> [1] 1.30700084350011
#> 
#> $UmbralRadii
#> [1] 0.71809915649989
#> 
#> $PenumbralMagnitude
#> [1] 0.653710092460876
#> 
#> $UmbralMagnitude
#> [1] -0.426843461667968
#> 
#> $PartialPhaseSemiDuration
#> [1] 0
#> 
#> $TotalPhaseSemiDuration
#> [1] 0
#> 
#> $PartialPhasePenumbraSemiDuration
#> [1] 115.585572595986
#>