CAAGalileanMoons_Calculate

CAAGalileanMoons_Calculate

CAAGalileanMoons_Calculate(JD)

Arguments

JD
The date in Dynamical time to calculate for.

Details

Value

A class which itself contains a class for each moon. This contained class itself contains

MeanLongitude The mean longitude of the moon in degrees.

TrueLongitude The true longitude of the moon in degrees.

TropicalLongitude The tropical longitude of the moon in degrees.

EquatorialLatitude The latitude in degrees of the moon with respect to Jupiter's equatorial plane.

r The radius vector of the moon in equatorial radii of Jupiter.

TrueRectangularCoordinates The true 3D rectangular coordinates of the moon.

ApparentRectangularCoordinates The apparent 3D rectangular coordinates of the moon.

bInTransit A Boolean which if true means that the moon is in front of Jupiter as viewed from the Earth otherwise false.

bInOccultation A Boolean which if true means that the moon is behind Jupiter as viewed from the Earth otherwise false.

bInEclipse A Boolean which if true means that the moon is behind Jupiter as viewed from the Sun otherwise false.

bInShadowTransit A Boolean which if true means that the moon is in front of Jupiter as viewed from the Earth otherwise false.

References

Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.

Author

C++ code by PJ Naughter, imported to R by Jinlong Zhang

Note

See also

Examples

CAAGalileanMoons_Calculate(2448972.50068);

#> $Longitude
#> $Longitude$Satellite1.MeanLongitude
#> [1] 335.606267330237
#> 
#> $Longitude$Satellite1.TrueLongitude
#> [1] 335.59973573708
#> 
#> $Longitude$Satellite2.MeanLongitude
#> [1] 62.3421151589137
#> 
#> $Longitude$Satellite2.TrueLongitude
#> [1] 63.4422246339964
#> 
#> $Longitude$Satellite3.MeanLongitude
#> [1] 15.7100460872171
#> 
#> $Longitude$Satellite3.TrueLongitude
#> [1] 15.7504351398675
#> 
#> $Longitude$Satellite4.MeanLongitude
#> [1] 26.1910398262262
#> 
#> $Longitude$Satellite4.TrueLongitude
#> [1] 26.782078134318
#> 
#> 
#> $EquatorialLatitude
#> $EquatorialLatitude$Satellite1.EquatorialLatitude
#> [1] 0.043408227809024
#> 
#> $EquatorialLatitude$Satellite2.EquatorialLatitude
#> [1] 0.156540424316933
#> 
#> $EquatorialLatitude$Satellite3.EquatorialLatitude
#> [1] -0.225526371384577
#> 
#> $EquatorialLatitude$Satellite4.EquatorialLatitude
#> [1] -0.148129134854337
#> 
#> 
#> $r
#> $r$Satellite1.r
#> [1] 5.92989277530137
#> 
#> $r$Satellite2.r
#> [1] 9.40373534633906
#> 
#> $r$Satellite3.r
#> [1] 15.0001974294798
#> 
#> $r$Satellite4.r
#> [1] 26.2129214736719
#> 
#> 
#> $TropicalLongitude
#> $TropicalLongitude$Satellite1.TropicalLongitude
#> [1] 336.199758351702
#> 
#> $TropicalLongitude$Satellite2.TropicalLongitude
#> [1] 64.0422472486184
#> 
#> $TropicalLongitude$Satellite3.TropicalLongitude
#> [1] 16.3504577544895
#> 
#> $TropicalLongitude$Satellite4.TropicalLongitude
#> [1] 27.38210074894
#> 
#> 
#> $TrueRectangularCoordinates
#> $TrueRectangularCoordinates$Satellite1.TrueRectangularCoordinates.X
#> [1] -3.44908139105702
#> 
#> $TrueRectangularCoordinates$Satellite1.TrueRectangularCoordinates.Y
#> [1] 0.213620176734819
#> 
#> $TrueRectangularCoordinates$Satellite1.TrueRectangularCoordinates.Z
#> [1] -4.81890364134077
#> 
#> $TrueRectangularCoordinates$Satellite2.TrueRectangularCoordinates.X
#> [1] 7.43804782241773
#> 
#> $TrueRectangularCoordinates$Satellite2.TrueRectangularCoordinates.Y
#> [1] 0.275123466951539
#> 
#> $TrueRectangularCoordinates$Satellite2.TrueRectangularCoordinates.Z
#> [1] -5.74717235980843
#> 
#> $TrueRectangularCoordinates$Satellite3.TrueRectangularCoordinates.X
#> [1] 1.19894363001318
#> 
#> $TrueRectangularCoordinates$Satellite4.TrueRectangularCoordinates.Y
#> [1] 1.02681812789619
#> 
#> $TrueRectangularCoordinates$Satellite4.TrueRectangularCoordinates.Z
#> [1] -25.2244342691323
#> 
#> 
#> $ApparentRectangularCoordinates
#> $ApparentRectangularCoordinates$Satellite1.ApparentRectangularCoordinates.X
#> [1] -3.4502566252194
#> 
#> $ApparentRectangularCoordinates$Satellite1.ApparentRectangularCoordinates.Y
#> [1] 0.213707008662089
#> 
#> $ApparentRectangularCoordinates$Satellite1.ApparentRectangularCoordinates.Z
#> [1] -4.81890364134077
#> 
#> $ApparentRectangularCoordinates$Satellite2.ApparentRectangularCoordinates.X
#> [1] 7.44181515173665
#> 
#> $ApparentRectangularCoordinates$Satellite2.ApparentRectangularCoordinates.Y
#> [1] 0.275256851275636
#> 
#> $ApparentRectangularCoordinates$Satellite2.ApparentRectangularCoordinates.Z
#> [1] -5.74717235980843
#> 
#> $ApparentRectangularCoordinates$Satellite3.ApparentRectangularCoordinates.X
#> [1] 1.20099698751054
#> 
#> $ApparentRectangularCoordinates$Satellite3.ApparentRectangularCoordinates.Y
#> [1] 0.590013116223285
#> 
#> $ApparentRectangularCoordinates$Satellite3.ApparentRectangularCoordinates.Z
#> [1] -14.9405896180217
#> 
#> $ApparentRectangularCoordinates$Satellite4.ApparentRectangularCoordinates.X
#> [1] 7.07196612342878
#> 
#> $ApparentRectangularCoordinates$Satellite4.ApparentRectangularCoordinates.Y
#> [1] 1.02900665561405
#> 
#> $ApparentRectangularCoordinates$Satellite4.ApparentRectangularCoordinates.Z
#> [1] -25.2244342691323
#> 
#> 
#> $Eclipse
#> $Eclipse$Satellite1.bInEclipse
#> [1] FALSE
#> 
#> $Eclipse$Satellite2.bInEclipse
#> [1] FALSE
#> 
#> $Eclipse$Satellite3.bInEclipse
#> [1] FALSE
#> 
#> $Eclipse$Satellite4.bInEclipse
#> [1] FALSE
#> 
#> $Eclipse$Satellite1.bInShadowTransit
#> [1] FALSE
#> 
#> $Eclipse$Satellite2.bInShadowTransit
#> [1] FALSE
#> 
#> $Eclipse$Satellite3.bInShadowTransit
#> [1] FALSE
#> 
#> $Eclipse$Satellite4.bInShadowTransit
#> [1] FALSE
#> 
#>