CAAParabolic_Calculate

CAAParabolic_Calculate

CAAParabolic_Calculate(JD, elements_JDEquinox, elements_omega, elements_w, elements_i, elements_T, elements_q)

Arguments

JD
JD The date in Dynamical time to calculate for.

elements_JDEquinox
JDEquinox The Julian Date for which equatorial coordinates should be calculated for.

elements_omega
omega The longitude of the ascending node in degrees.

elements_w
w The argument of the perihelion in degrees.

elements_i
i The inclination of the plane of the orbit in degrees.

elements_T
T The Julian date of the time of passage in perihelion.

elements_q
q The perihelion distance in astronomical units.

Details

Value

A class containing

HeliocentricRectangularEquatorial 3D rectangular heliocentric equatorial coordinates of the object.

HeliocentricRectangularEcliptical 3D rectangular heliocentric ecliptical coordinates of the object.

HeliocentricEclipticLongitude The heliocentric ecliptical longitude in degrees of the object.

HeliocentricEclipticLatitude The heliocentric ecliptical latitude in degrees of the object.

TrueGeocentricRA The geocentric right ascension of the object as an hour angle (i.e. without light time correction applied).

TrueGeocentricDeclination The geocentric declination of the object in degrees (i.e. without light time correction applied).

TrueGeocentricDistance The true distance in astronomical units between the Earth and the object.

TrueGeocentricLightTime The light travel time in days from the Earth to the object.

AstrometricGeocenticRA The geocentric right ascension of the object as an hour angle (i.e. with light time correction applied)

AstrometricGeocentricDeclination The geocentric declination of the object in degrees (i.e. with light time correction applied)

AstrometricGeocentricDistance The observed distance of the object in astronomical units.

AstrometricGeocentricLightTime The observed light travel time of the object in days.

Elongation The elongation of the object to the Sun in degrees.

PhaseAngle The phase angle (the angle Sun - object - Earth) in degrees.

References

Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.

Author

C++ code by PJ Naughter, imported to R by Jinlong Zhang

Note

See also

Examples

CAAParabolic_Calculate(JD = 2456597.5, elements_JDEquinox =2451544.5, 
elements_omega = 22, elements_w = 39, 
elements_i = 11, elements_T = 2456521.54432212, elements_q = 5.53)

#> $HeliocentricRectangularEquatorial.X
#> [1] 2.00944046318696
#> 
#> $HeliocentricRectangularEquatorial.Y
#> [1] 4.39141116659296
#> 
#> $HeliocentricRectangularEquatorial.Z
#> [1] 2.75082051643159
#> 
#> $HeliocentricRectangularEcliptical.X
#> [1] 2.00944046318696
#> 
#> $HeliocentricRectangularEcliptical.Y
#> [1] 5.12325453643857
#> 
#> $HeliocentricRectangularEcliptical.Z
#> [1] 0.777025419506919
#> 
#> $HeliocentricEclipticLongitude
#> [1] 68.5839188297583
#> 
#> $HeliocentricEclipticLatitude
#> [1] 8.03671333128148
#> 
#> $TrueGeocentricRA
#> [1] 4.80859149289289
#> 
#> $TrueGeocentricDeclination
#> [1] 31.9403321399556
#> 
#> $TrueGeocentricDistance
#> [1] 4.73476363840754
#> 
#> $TrueGeocentricLightTime
#> [1] 0.0273457140397973
#> 
#> $AstrometricGeocenticRA
#> [1] 4.80833492315627
#> 
#> $AstrometricGeocentricDeclination
#> [1] 31.9395400405936
#> 
#> $AstrometricGeocentricDistance
#> [1] 4.73471356601553
#> 
#> $AstrometricGeocentricLightTime
#> [1] 0.0273454248457809
#> 
#> $Elongation
#> [1] 142.795467307919
#> 
#> $PhaseAngle
#> [1] 6.1994373694602
#>