CAASaturnMoons_Calculate

CAASaturnMoons_Calculate

CAASaturnMoons_Calculate(JD)

Arguments

JD
The date in Dynamical time to calculate for.

Details

Value

A class which itself contains a class for each moon. This contained class itself contains

TrueRectangularCoordinates The true 3D rectangular coordinates of the moon.

ApparentRectangularCoordinates The apparent 3D rectangular coordinates of the moon.

bInTransit A Boolean which if true means that the moon is in front of Saturn as viewed from the Earth otherwise false.

bInOccultation A Boolean which if true means that the moon is behind Saturn as viewed from the Earth otherwise false.

bInEclipse A Boolean which if true means that the moon is behind Saturn as viewed from the Sun otherwise false.

bInShadowTransit A Boolean which if true means that the moon is in front of Saturn as viewed from the Earth otherwise false.

References

Meeus, J. H. (1991). Astronomical algorithms. Willmann-Bell, Incorporated.

Author

C++ code by PJ Naughter, imported to R by Jinlong Zhang

Note

See also

Examples

JD = 2456597.5
CAASaturnMoons_Calculate(JD)

#> $Satellite1
#> $Satellite1$Satellite1.TrueRectangularCoordinates.X
#> [1] -3.01541345894479
#> 
#> $Satellite1$Satellite1.TrueRectangularCoordinates.Y
#> [1] 0.0255344971629219
#> 
#> $Satellite1$Satellite1.TrueRectangularCoordinates.Z
#> [1] -0.00241402893824161
#> 
#> $Satellite1$Satellite1.ApparentRectangularCoordinates.X
#> [1] -3.01541372891928
#> 
#> $Satellite1$Satellite1.ApparentRectangularCoordinates.Y
#> [1] 0.0255344994573641
#> 
#> $Satellite1$Satellite1.ApparentRectangularCoordinates.Z
#> [1] -0.00241402893824161
#> 
#> $Satellite1$Satellite1.bInTransit
#> [1] FALSE
#> 
#> $Satellite1$Satellite1.bInOccultation
#> [1] FALSE
#> 
#> $Satellite1$Satellite1.bInEclipse
#> [1] FALSE
#> 
#> $Satellite1$Satellite1.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite2
#> $Satellite2$Satellite2.TrueRectangularCoordinates.X
#> [1] 1.83111227471239
#> 
#> $Satellite2$Satellite2.TrueRectangularCoordinates.Y
#> [1] -1.21928160558093
#> 
#> $Satellite2$Satellite2.TrueRectangularCoordinates.Z
#> [1] -3.28788205427744
#> 
#> $Satellite2$Satellite2.ApparentRectangularCoordinates.X
#> [1] 1.83145931073649
#> 
#> $Satellite2$Satellite2.ApparentRectangularCoordinates.Y
#> [1] -1.21943084402376
#> 
#> $Satellite2$Satellite2.ApparentRectangularCoordinates.Z
#> [1] -3.28788205427744
#> 
#> $Satellite2$Satellite2.bInTransit
#> [1] FALSE
#> 
#> $Satellite2$Satellite2.bInOccultation
#> [1] FALSE
#> 
#> $Satellite2$Satellite2.bInEclipse
#> [1] FALSE
#> 
#> $Satellite2$Satellite2.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite3
#> $Satellite3$Satellite3.TrueRectangularCoordinates.X
#> [1] 3.61073175291721
#> 
#> $Satellite3$Satellite3.TrueRectangularCoordinates.Y
#> [1] -1.17500156226089
#> 
#> $Satellite3$Satellite3.TrueRectangularCoordinates.Z
#> [1] -3.06694131231164
#> 
#> $Satellite3$Satellite3.ApparentRectangularCoordinates.X
#> [1] 3.61122223193694
#> 
#> $Satellite3$Satellite3.ApparentRectangularCoordinates.Y
#> [1] -1.17513571538622
#> 
#> $Satellite3$Satellite3.ApparentRectangularCoordinates.Z
#> [1] -3.06694131231164
#> 
#> $Satellite3$Satellite3.bInTransit
#> [1] FALSE
#> 
#> $Satellite3$Satellite3.bInOccultation
#> [1] FALSE
#> 
#> $Satellite3$Satellite3.bInEclipse
#> [1] FALSE
#> 
#> $Satellite3$Satellite3.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite4
#> $Satellite4$Satellite4.TrueRectangularCoordinates.X
#> [1] -6.19519479923219
#> 
#> $Satellite4$Satellite4.TrueRectangularCoordinates.Y
#> [1] -0.289077916690915
#> 
#> $Satellite4$Satellite4.TrueRectangularCoordinates.Z
#> [1] -0.776430654237804
#> 
#> $Satellite4$Satellite4.ApparentRectangularCoordinates.X
#> [1] -6.19537040560768
#> 
#> $Satellite4$Satellite4.ApparentRectangularCoordinates.Y
#> [1] -0.28908627151759
#> 
#> $Satellite4$Satellite4.ApparentRectangularCoordinates.Z
#> [1] -0.776430654237804
#> 
#> $Satellite4$Satellite4.bInTransit
#> [1] FALSE
#> 
#> $Satellite4$Satellite4.bInOccultation
#> [1] FALSE
#> 
#> $Satellite4$Satellite4.bInEclipse
#> [1] FALSE
#> 
#> $Satellite4$Satellite4.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite5
#> $Satellite5$Satellite5.TrueRectangularCoordinates.X
#> [1] -8.18894419819549
#> 
#> $Satellite5$Satellite5.TrueRectangularCoordinates.Y
#> [1] -1.05229983672617
#> 
#> $Satellite5$Satellite5.TrueRectangularCoordinates.Z
#> [1] -2.80460687040632
#> 
#> $Satellite5$Satellite5.ApparentRectangularCoordinates.X
#> [1] -8.1897718844009
#> 
#> $Satellite5$Satellite5.ApparentRectangularCoordinates.Y
#> [1] -1.0524097029355
#> 
#> $Satellite5$Satellite5.ApparentRectangularCoordinates.Z
#> [1] -2.80460687040632
#> 
#> $Satellite5$Satellite5.bInTransit
#> [1] FALSE
#> 
#> $Satellite5$Satellite5.bInOccultation
#> [1] FALSE
#> 
#> $Satellite5$Satellite5.bInEclipse
#> [1] FALSE
#> 
#> $Satellite5$Satellite5.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite6
#> $Satellite6$Satellite6.TrueRectangularCoordinates.X
#> [1] 15.2711163368771
#> 
#> $Satellite6$Satellite6.TrueRectangularCoordinates.Y
#> [1] -4.56342716148108
#> 
#> $Satellite6$Satellite6.TrueRectangularCoordinates.Z
#> [1] -12.7113433524391
#> 
#> $Satellite6$Satellite6.ApparentRectangularCoordinates.X
#> [1] 15.2785018829976
#> 
#> $Satellite6$Satellite6.ApparentRectangularCoordinates.Y
#> [1] -4.56558736869089
#> 
#> $Satellite6$Satellite6.ApparentRectangularCoordinates.Z
#> [1] -12.7113433524391
#> 
#> $Satellite6$Satellite6.bInTransit
#> [1] FALSE
#> 
#> $Satellite6$Satellite6.bInOccultation
#> [1] FALSE
#> 
#> $Satellite6$Satellite6.bInEclipse
#> [1] FALSE
#> 
#> $Satellite6$Satellite6.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite7
#> $Satellite7$Satellite7.TrueRectangularCoordinates.X
#> [1] -8.52560546462376
#> 
#> $Satellite7$Satellite7.TrueRectangularCoordinates.Y
#> [1] 8.28626984476447
#> 
#> $Satellite7$Satellite7.TrueRectangularCoordinates.Z
#> [1] 23.3984164932289
#> 
#> $Satellite7$Satellite7.ApparentRectangularCoordinates.X
#> [1] -8.51781319551348
#> 
#> $Satellite7$Satellite7.ApparentRectangularCoordinates.Y
#> [1] 8.27905918908061
#> 
#> $Satellite7$Satellite7.ApparentRectangularCoordinates.Z
#> [1] 23.3984164932289
#> 
#> $Satellite7$Satellite7.bInTransit
#> [1] FALSE
#> 
#> $Satellite7$Satellite7.bInOccultation
#> [1] FALSE
#> 
#> $Satellite7$Satellite7.bInEclipse
#> [1] FALSE
#> 
#> $Satellite7$Satellite7.bInShadowTransit
#> [1] FALSE
#> 
#> 
#> $Satellite8
#> $Satellite8$Satellite8.TrueRectangularCoordinates.X
#> [1] 50.5022984453394
#> 
#> $Satellite8$Satellite8.TrueRectangularCoordinates.Y
#> [1] 5.75970424963346
#> 
#> $Satellite8$Satellite8.TrueRectangularCoordinates.Z
#> [1] -29.7795350452048
#> 
#> $Satellite8$Satellite8.ApparentRectangularCoordinates.X
#> [1] 50.5585081845957
#> 
#> $Satellite8$Satellite8.ApparentRectangularCoordinates.Y
#> [1] 5.76609581258491
#> 
#> $Satellite8$Satellite8.ApparentRectangularCoordinates.Z
#> [1] -29.7795350452048
#> 
#> $Satellite8$Satellite8.bInTransit
#> [1] FALSE
#> 
#> $Satellite8$Satellite8.bInOccultation
#> [1] FALSE
#> 
#> $Satellite8$Satellite8.bInEclipse
#> [1] FALSE
#> 
#> $Satellite8$Satellite8.bInShadowTransit
#> [1] FALSE
#> 
#>